refactor: Trial 분리 - Trial 클래스를 분리하고 GameSetter 삭제 - 관련 메서드 및 테스트 수정
